backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hackathon-todo-backend"
version = "0.1.0"
description = "FastAPI backend for Hackathon Todo application"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.8",
    "uvicorn[standard]==0.34.0",
    "sqlmodel==0.0.22",
    "asyncpg==0.30.0",
    "sqlalchemy[asyncio]==2.0.37",
    "alembic==1.14.1",
    "httpx==0.28.1",
    "pydantic-settings==2.7.1",
    "greenlet==3.1.1",
    "openai-agents>=0.0.16",
    "openai-chatkit>=0.0.3",
    "mcp[cli]>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.3",
    "httpx==0.28.1",
]

[tool.setuptools.packages.find]
include = ["app*"]
